WormGPT_Base/requirements.txt
streamlit
google-generativeai
cryptography